Added a check for the maxauthors processing. This is currently working, so this is a check to make sure it doesn't break in future.
I notice that the \bibitem optional argument could potentially be very long, if
there are lots of authors.  This is in principle a problem, since it could get
out of control if there are dozens or even hundreds of authors, but the excess
names don't appear in the formatted output, so I don't propose to change this
until an actual problem emerges.